feat: Implemented account deletion and refactored settings components
